requirements.txt
anyio==3.6.1
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.81.0
Flask==2.2.2
Flask-Compress==1.12
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
orjson==3.7.12
pandas==1.4.3
plotly==5.10.0
pydantic==1.9.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tenacity==8.0.1
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.2.2
zipp==3.8.1
